requirements.txt
# Core ML dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
tokenizers>=0.13.0

# Web interface
gradio>=3.35.0
streamlit>=1.24.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0

# Enhanced RAG dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# System monitoring
psutil>=5.9.0

# Metrics
sacrebleu>=2.4.0
rouge-score>=0.1.2
bert-score>=0.3.13
# Optional: for CodeBLEU metric; fallback in code if missing
codebleu>=0.6.0

# Development tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Logging and monitoring  
wandb>=0.15.0
tensorboard>=2.13.0
rich>=13.0.0

# Additional utilities
click>=8.1.0
tqdm>=4.65.0

# Security and validation
ast>=3.9.0

# HTTP requests (for security testing)
requests>=2.31.0
requests>=2.31.0